Bump version codes and update pricing feature
- Updated Android versionCode from 428 to 429 and iOS CURRENT_PROJECT_VERSION from 445 to 446.
- Modified Info.plist to reflect the new version number.
- Enhanced the KalemEkleView to conditionally display the latest price based on user permissions and fetch price summary data.
- Introduced new observable properties and actions in KalemEkleViewModel to manage stock price summaries.
- Updated StokFiyatOzetiModel to use double instead of int for price-related fields for better precision.
- Adjusted JSON serialization methods to accommodate the new data types.
- Updated web index.html to use the latest flutter.js version 244.

+++ lib/view/main_page/alt_sayfalar/stok/fiyat_ozeti/model/stok_fiyat_ozeti_model.dart
@@ -15,11 +15,11 @@ class StokFiyatOzetiModel with _$StokFiyatOzetiModel, NetworkManagerMixin {
     String? cariAdi,
     DateTime? tarih,
     double? fiyat,
-    int? dovizFiyati,
+    double? dovizFiyati,
     String? dovizAdi,
     int? dovizTipi,
-    int? brutFiyat,
-    int? brutDovizFiyati,
+    double? brutFiyat,
+    double? brutDovizFiyati,
   }) = _StokFiyatOzetiModel;
   const StokFiyatOzetiModel._();
